pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "requests>=2.32.3",
]
